pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "marketplace-intelligence-platform"
version = "0.1.0"
description = "Marketplace Intelligence Platform – Graph-powered marketplace analytics"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = "MIT"
authors = [{ name = "Evgenii Saurov" }]

dependencies = [
  "fastapi>=0.120.1",
  "fastembed>=0.7.3",
  "minio>=7.2.18",
  "neo4j>=5.28.2",
  "openai>=2.6.1",
  "orjson>=3.11.4",
  "pydantic>=2.9,<3",
  "python-dotenv>=1.0,<2",
  "python-ulid>=3.1.0",
  "qdrant-client>=1.15.1",
  "redis>=5.3.1",
  "rq>=2.6.0",
  "uvicorn>=0.38.0",
]

[project.optional-dependencies]
api = [
   "neo4j>=5.21,<6",
   "minio>=7,<8",
   "pandas>=2.2,<3",
   "qdrant-client>=1.9",
   "fastembed>=0.3.2",
   "httpx>=0.27,<1",
   "openpyxl>=3.1,<4",
   "python-multipart>=0.0.9",
   # European AI & integrations
   "mistralai>=1.0,<2",  # French LLM provider (Paris 🇫🇷)
]
etl = [
  "neo4j>=5.21,<6",
  "minio>=7,<8",
  "pandas>=2.2,<3",
]
ingestors = [
  "minio>=7,<8",
  "requests>=2.32,<3",
]
dev = [
  "ipython>=8.20,<9",
  "ruff>=0.6,<1",
  "black>=24.8,<25",
  "mypy>=1.10,<2",
]
test = [
  "pytest>=8.2,<9",
  "pytest-asyncio>=0.23,<1",
  "httpx>=0.27,<1",
  "respx>=0.21,<1",
  "requests-mock>=1.12,<2",
]

[project.scripts]
atlas-apollo-fetch = "atlas.ingestors.apollo.apollo_fetch:main"
atlas-etl-apollo = "atlas.etl.apollo_to_graph.etl_apollo:main"
atlas-api = "atlas.services.query_api.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/atlas"]

[tool.uv]
python-preference = "only-managed"

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E","F","I","B","UP","SIM"]
lint.ignore = ["E501"]
exclude = [".venv", "infra", "anthropic", "tools"]

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = "\\.venv|infra|anthropic|tools"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests", "src/atlas"]
addopts = "-q"
pythonpath = ["src"]   # <— make imports work under pytest

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
exclude = "(.venv|infra|tools|anthropic|docker)"
